XMLHttpRequest/status-values.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>XMLHttpRequest Test: status - values</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<meta name="assert" content="Check if XMLHttpRequest.status return 0 if state is UNSENT or OPENED">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>

<script>

async_test(function (t) {
  var client = new XMLHttpRequest();
  if (client.readyState == client.UNSENT) {
    t.step(function () {
      assert_equals(client.status, 0);
    });
  }

  client.onreadystatechange = t.step_func(function () {
    if (client.readyState == client.OPENED) {
      assert_equals(client.status, 0);
      t.done();
    }
  });

  client.open("GET", "resources/headers.py")
}, document.title);

</script>
